+++ src/conformly/constraints/__init__.py
@@ -1,4 +1,5 @@
 from .base import Constraint
+from .enum import OneOf
 from .numeric import GreaterOrEqual, GreaterThan, LessOrEqual, LessThan
 from .string import MaxLength, MinLength, Pattern
 from .types import ConstraintType
@@ -12,5 +13,6 @@ __all__ = [
     "LessThan",
     "MaxLength",
     "MinLength",
+    "OneOf",
     "Pattern",
 ]